+++ rest_framework_services/types/__init__.py
@@ -20,6 +20,7 @@ from rest_framework_services.types.no_input import NoInput
 from rest_framework_services.types.offline_context import OfflineContext
 from rest_framework_services.types.offline_service_view import OfflineServiceView
 from rest_framework_services.types.polymorphic_service_spec import PolymorphicServiceSpec
+from rest_framework_services.types.registered_spec import RegisteredSpec
 from rest_framework_services.types.selector_kind import SelectorKind
 from rest_framework_services.types.selector_spec import SelectorSpec
 from rest_framework_services.types.service_spec import ServiceSpec
@@ -43,6 +44,7 @@ __all__ = [
     "OfflineContext",
     "OfflineServiceView",
     "PolymorphicServiceSpec",
+    "RegisteredSpec",
     "SelectorKind",
     "SelectorSpec",
     "ServiceSpec",